udpの送受信をintからlongに変更

+++ include/udp_connect.hpp
@@ -36,11 +36,11 @@ class UdpConnect{
 
         UdpConnect(std::string address, int port, size_t element_count); // UDPコンストラクタ       
         
-        void udp_send(const std::vector<double>& values , int roop_count); // UDP送信関数
+        void udp_send(const std::vector<double>& values , long roop_count); // UDP送信関数
 
         void udp_bind(); 
 
-        std::pair<std::vector<double>, int> udp_recv();
+        std::pair<std::vector<double>, long> udp_recv();
 
         
         ~UdpConnect();
@@ -52,8 +52,8 @@ class UdpCommunicator{
     * UDP communication processing class
     */
    public:
-        UdpCommunicator(std::deque<std::pair<std::vector<double>, int>>& deque_master, // address of deque_master
-            std::deque<std::pair<std::vector<double>, int>>& deque_copy,
+        UdpCommunicator(std::deque<std::pair<std::vector<double>, long>>& deque_master, // address of deque_master
+            std::deque<std::pair<std::vector<double>, long>>& deque_copy,
             std::mutex& queue_mutex_master,
             std::mutex& queue_mutex_copy);
 
@@ -66,8 +66,8 @@ class UdpCommunicator{
         ~UdpCommunicator();
 
     private:
-        std::deque<std::pair<std::vector<double>, int>>& deque_master_;
-        std::deque<std::pair<std::vector<double>, int>>& deque_copy_;
+        std::deque<std::pair<std::vector<double>, long>>& deque_master_;
+        std::deque<std::pair<std::vector<double>, long>>& deque_copy_;
         std::mutex& queue_mutex_master_;
         std::mutex& queue_mutex_copy_;
 };
